requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
attrs==19.3.0
Automat==20.2.0
bcrypt==3.1.7
bleach==1.5.0
cached-property==1.5.1
captcha==0.3
certifi==2019.11.28
cffi==1.14.0
Click==7.0
constantly==15.1.0
crypto==1.4.1
cryptography==2.9
cssselect==1.1.0
cycler==0.10.0
dnspython==2.0.0
dockerpty==0.4.1
docopt==0.6.2
eventlet==0.25.2
fabric==2.5.0
Flask==1.1.1
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-JSON==0.3.4
futures==3.1.1
gast==0.3.3
google-pasta==0.2.0
GPUtil==1.4.0
graphviz==0.14.1
greenlet==0.4.16
grpcio==1.28.0
h5py==2.10.0
html5lib==0.9999999
hyperlink==19.0.0
idna==2.7
importlib-metadata==1.6.0
incremental==17.5.0
invoke==1.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
Keras==2.1.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mock==4.0.2
monotonic==1.5
Naked==0.1.31
numpy==1.16.0
paramiko==2.7.1
parsel==1.5.2
pika==1.1.0
Pillow==7.2.0
Protego==0.1.16
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyDispatcher==2.0.5
pydot==1.4.1
PyHamcrest==2.0.2
pymongo==3.6.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
queuelib==1.5.0
redis==3.4.1
requests==2.23.0
scipy==1.5.2
Scrapy==2.0.1
scrapy-redis==0.6.8
service-identity==18.1.0
shellescape==3.8.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.11.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
texttable==1.6.2
tornado==4.3
Twisted==20.3.0
urllib3==1.23
w3lib==1.21.0
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
